Fix unfair verification in "override" problem; revert Number patch from e639d8a85f0f120965638b783978b4f9be6df831.
In the earlier "client" exercise you are told to transform the
process.argv argument into numbers. This suggests doing it in the client
(rather than in the plugin); and in fact the solution does just that.

As such, our plugin code expects the service client to take care of the
transformation, which leads to a problem in the "override" exercise,
where the verifier suddenly supplies number strings. The original
solution was to patch up the sum function and that's probably not a bad
idea in general, but it distracts from what's being taught. This patch
has the verifier use either (non-numerical) strings or proper numbers.